# out/src/Johnfg2610.Stockit/wwwroot/openapi-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "contact" : {
      "email" : "johnfg2610@gmail.com",
      "name" : "john",
      "url" : "https://www.github.com/johnfg2610/stockit"
    },
    "description" : "Used to access inventory counts as well as brands and catalog items",
    "license" : {
      "name" : "ISC",
      "url" : "https://github.com/johnfg2610/stockit/blob/master/LICENSE"
    },
    "title" : "Inventory",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:3000"
  } ],
  "security" : [ {
    "jwt-auth" : [ ]
  } ],
...
  "paths" : {
...
    "/brand" : {
      "post" : {
        "description" : "Used to create a new brand",
        "operationId" : "post-brand",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "examples" : {
                "example-1" : {
                  "value" : {
                    "name" : "string",
                    "description" : "string",
                    "iconUrl" : "string"
                  }
                }
              },
              "schema" : {
                "$ref" : "#/components/schemas/BrandForm"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Not Found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Create a new Brand",
        "tags" : [ "brand" ]
      }
    },
    "/brand/{brandId}" : {
      "delete" : {
        "description" : "Delete a brand by ID",
        "operationId" : "delete-brand-brandId",
        "parameters" : [ {
          "description" : "the id of the brand to access",
          "explode" : false,
          "in" : "path",
          "name" : "brandId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "201" : {
            "description" : "Created"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Not Found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Delete a brand by ID",
        "tags" : [ "brand" ]
      },
      "get" : {
        "description" : "Get a brand by ID",
        "operationId" : "get-brand-brandId",
        "parameters" : [ {
          "description" : "the id of the brand to access",
          "explode" : false,
          "in" : "path",
          "name" : "brandId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "example-1" : {
                    "value" : {
                      "id" : "string",
                      "name" : "string",
                      "description" : "string",
                      "iconUrl" : "string"
                    }
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/Brand"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Not Found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Get brand by ID",
        "tags" : [ "brand" ]
      },
      "put" : {
        "description" : "Replaces a brand by ID",
        "operationId" : "put-brand-brandId",
        "parameters" : [ {
          "description" : "the id of the brand to access",
          "explode" : false,
          "in" : "path",
          "name" : "brandId",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BrandForm"
              }
            }
          },
          "description" : "The brand object to replace this with"
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Not Found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Replace a brand by ID",
        "tags" : [ "brand" ]
      }
    },
...
  },
  "components" : {
...
    "schemas" : {
...
      "Brand" : {
        "description" : "A brand is used to display who produces a item, one brand can be used by many sellers",
        "example" : {
          "name" : "name",
          "description" : "description",
          "id" : "id",
          "iconUrl" : "iconUrl"
        },
        "properties" : {
          "id" : {
            "description" : "a random identifier to identify this brand",
            "type" : "string"
          },
          "name" : {
            "description" : "the name of this brand",
            "type" : "string"
          },
          "description" : {
            "description" : "the description to discribe this brand",
            "type" : "string"
          },
          "iconUrl" : {
            "description" : "a optional url pointing to a icon url",
            "type" : "string"
          }
        },
        "required" : [ "description", "name" ],
        "title" : "Brand",
        "type" : "object",
        "x-tags" : [ "brand" ]
      },
      "BrandForm" : {
        "description" : "A brand is used to display who produces a item, one brand can be used by many sellers",
        "example" : {
          "name" : "name",
          "description" : "description",
          "iconUrl" : "iconUrl"
        },
        "properties" : {
          "name" : {
            "description" : "the name of this brand",
            "type" : "string"
          },
          "description" : {
            "description" : "the description to discribe this brand",
            "type" : "string"
          },
          "iconUrl" : {
            "description" : "a optional url pointing to a icon url",
            "type" : "string"
          }
        },
        "required" : [ "description", "name" ],
        "title" : "BrandForm",
        "type" : "object",
        "x-tags" : [ "brand" ]
      },
      "Error" : {
        "description" : "Used to describe a error that happened during the execution of a service",
        "properties" : {
          "code" : {
            "type" : "number"
          },
          "message" : {
            "type" : "string"
          },
          "requestId" : {
            "type" : "string"
          }
        },
        "title" : "Error",
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "jwt-auth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}